+++ scripts/bundles/dev-server.ts
@@ -4,7 +4,6 @@ import { dataToEsm } from '@rollup/pluginutils';
 import fs from 'fs-extra';
 import { join } from 'path';
 import type { OutputChunk, Plugin, RollupOptions } from 'rollup';
-import { minify } from 'terser';
 import ts from 'typescript';
 
 import { getBanner } from '../utils/banner';
@@ -182,6 +181,8 @@ export async function devServer(opts: BuildOptions) {
             code = tsResults.outputText;
 
             code = intro + code + outro;
+
+            const { minify } = await import('terser');
 
             if (opts.isProd) {
               const minifyResults = await minify(code, {